+++ packages/toolpad-components/tsconfig.json
@@ -5,8 +5,7 @@
       // TODO: remove when typescript module resolution supports package exports. (v5.0 with module: 'bundler' ?)
       "@mui/toolpad-core": ["../toolpad-core/dist/esm"],
       "@mui/toolpad-core/path": ["../toolpad-core/dist/esm/path"],
-      "@mui/toolpad-core/objectKey": ["../toolpad-core/dist/esm/objectKey"],
-      "@mui/toolpad-core/utils/*": ["../toolpad-core/dist/esm/utils/*"]
+      "@mui/toolpad-core/objectKey": ["../toolpad-core/dist/esm/objectKey"]
     },
     "module": "esnext",
     "alwaysStrict": true,
